+++ CHANGELOG.md
@@ -10,10 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Made `Calendar` and `Options` `Eq`
 - Added function `Calendar::title`
 - Added week numbers
+- Added festivals
 
 ### Changed
 - Made `Calendar` and `YearCalendar` limited
 - Updated translations
+- Changed functions in `chinese_date.rs` to use `ChineseYear` and `ChineseMonth` as parameters
 
 ## [0.3.1] - 2024-12-25
 ### Fixed